+++ .gitignore
@@ -7,7 +7,7 @@ CMakeSettings.json
 # CMake configured files
 cpp/open3d/Open3D.h
 cpp/open3d/Open3DConfig.h
-cpp/open3d/visualization/Shader/Shader.h
+cpp/open3d/visualization/shader/Shader.h
 util/pip_package/setup.py
 
 # CMake temp file from Visual Studio
