+++ cpp/open3d/visualization/shader/GLSL/ImageFragmentShader.glsl
