+++ cpp/open3d/visualization/shader/GLSL/ImageMaskFragmentShader.glsl
