+++ cpp/open3d/visualization/shader/GLSL/ImageMaskVertexShader.glsl
