+++ cpp/open3d/visualization/shader/GLSL/ImageVertexShader.glsl
