+++ cpp/open3d/visualization/shader/GLSL/NormalFragmentShader.glsl
