+++ cpp/open3d/visualization/shader/GLSL/NormalVertexShader.glsl
