+++ cpp/open3d/visualization/shader/GLSL/PhongFragmentShader.glsl
