+++ cpp/open3d/visualization/shader/GLSL/PhongVertexShader.glsl
