+++ cpp/open3d/visualization/shader/GLSL/PickingFragmentShader.glsl
