+++ cpp/open3d/visualization/shader/GLSL/PickingVertexShader.glsl
