+++ cpp/open3d/visualization/shader/GLSL/RGBDImageFragmentShader.glsl
