+++ cpp/open3d/visualization/shader/GLSL/Simple2DFragmentShader.glsl
