+++ cpp/open3d/visualization/shader/GLSL/SimpleBlackFragmentShader.glsl
