+++ cpp/open3d/visualization/shader/GLSL/SimpleBlackVertexShader.glsl
