+++ cpp/open3d/visualization/shader/GLSL/SimpleFragmentShader.glsl
