+++ cpp/open3d/visualization/shader/GLSL/SimpleVertexShader.glsl
