+++ cpp/open3d/visualization/shader/GLSL/TexturePhongFragmentShader.glsl
