+++ cpp/open3d/visualization/shader/GLSL/TexturePhongVertexShader.glsl
