+++ cpp/open3d/visualization/shader/GLSL/TextureSimpleFragmentShader.glsl
