+++ cpp/open3d/visualization/shader/GLSL/TextureSimpleVertexShader.glsl
