+++ cpp/open3d/visualization/shader/ImageShader.h
@@ -26,24 +26,20 @@
 
 #pragma once
 
-#include <Eigen/Core>
-#include <vector>
-
-#include "open3d/visualization/Shader/ShaderWrapper.h"
+#include "open3d/geometry/Image.h"
+#include "open3d/visualization/shader/ShaderWrapper.h"
 
 namespace open3d {
 namespace visualization {
 
 namespace glsl {
 
-class TexturePhongShader : public ShaderWrapper {
+class ImageShader : public ShaderWrapper {
 public:
-    ~TexturePhongShader() override { Release(); }
+    ~ImageShader() override { Release(); }
 
 protected:
-    TexturePhongShader(const std::string &name) : ShaderWrapper(name) {
-        Compile();
-    }
+    ImageShader(const std::string &name) : ShaderWrapper(name) { Compile(); }
 
 protected:
     bool Compile() final;
@@ -63,62 +59,32 @@ protected:
     virtual bool PrepareBinding(const geometry::Geometry &geometry,
                                 const RenderOption &option,
                                 const ViewControl &view,
-                                std::vector<Eigen::Vector3f> &points,
-                                std::vector<Eigen::Vector3f> &normals,
-                                std::vector<Eigen::Vector2f> &uvs) = 0;
-
-protected:
-    void SetLighting(const ViewControl &view, const RenderOption &option);
+                                geometry::Image &image) = 0;
 
 protected:
     GLuint vertex_position_;
-    GLuint vertex_uv_;
-    GLuint vertex_normal_;
-    GLuint MVP_;
-    GLuint V_;
-    GLuint M_;
-    GLuint light_position_world_;
-    GLuint light_color_;
-    GLuint light_diffuse_power_;
-    GLuint light_specular_power_;
-    GLuint light_specular_shininess_;
-    GLuint light_ambient_;
+    GLuint vertex_position_buffer_;
+    GLuint vertex_UV_;
+    GLuint vertex_UV_buffer_;
+    GLuint image_texture_;
+    GLuint image_texture_buffer_;
+    GLuint vertex_scale_;
 
-    GLuint diffuse_texture_;
-
-    int num_materials_;
-    std::vector<int> array_offsets_;
-    std::vector<GLsizei> draw_array_sizes_;
-
-    std::vector<GLuint> vertex_position_buffers_;
-    std::vector<GLuint> vertex_uv_buffers_;
-    std::vector<GLuint> vertex_normal_buffers_;
-    std::vector<GLuint> diffuse_texture_buffers_;
-
-    // At most support 4 lights
-    GLHelper::GLMatrix4f light_position_world_data_;
-    GLHelper::GLMatrix4f light_color_data_;
-    GLHelper::GLVector4f light_diffuse_power_data_;
-    GLHelper::GLVector4f light_specular_power_data_;
-    GLHelper::GLVector4f light_specular_shininess_data_;
-    GLHelper::GLVector4f light_ambient_data_;
+    GLHelper::GLVector3f vertex_scale_data_;
 };
 
-class TexturePhongShaderForTriangleMesh : public TexturePhongShader {
+class ImageShaderForImage : public ImageShader {
 public:
-    TexturePhongShaderForTriangleMesh()
-        : TexturePhongShader("TexturePhongShaderForTriangleMesh") {}
+    ImageShaderForImage() : ImageShader("ImageShaderForImage") {}
 
 protected:
-    bool PrepareRendering(const geometry::Geometry &geometry,
-                          const RenderOption &option,
-                          const ViewControl &view) final;
-    bool PrepareBinding(const geometry::Geometry &geometry,
-                        const RenderOption &option,
-                        const ViewControl &view,
-                        std::vector<Eigen::Vector3f> &points,
-                        std::vector<Eigen::Vector3f> &normals,
-                        std::vector<Eigen::Vector2f> &uvs) final;
+    virtual bool PrepareRendering(const geometry::Geometry &geometry,
+                                  const RenderOption &option,
+                                  const ViewControl &view) final;
+    virtual bool PrepareBinding(const geometry::Geometry &geometry,
+                                const RenderOption &option,
+                                const ViewControl &view,
+                                geometry::Image &render_image) final;
 };
 
 }  // namespace glsl
