+++ cpp/open3d/visualization/shader/PhongShader.h
@@ -29,21 +29,19 @@
 #include <Eigen/Core>
 #include <vector>
 
-#include "open3d/visualization/Shader/ShaderWrapper.h"
+#include "open3d/visualization/shader/ShaderWrapper.h"
 
 namespace open3d {
 namespace visualization {
 
 namespace glsl {
 
-class TextureSimpleShader : public ShaderWrapper {
+class PhongShader : public ShaderWrapper {
 public:
-    ~TextureSimpleShader() override { Release(); }
+    ~PhongShader() override { Release(); }
 
 protected:
-    TextureSimpleShader(const std::string &name) : ShaderWrapper(name) {
-        Compile();
-    }
+    PhongShader(const std::string &name) : ShaderWrapper(name) { Compile(); }
 
 protected:
     bool Compile() final;
@@ -64,27 +62,41 @@ protected:
                                 const RenderOption &option,
                                 const ViewControl &view,
                                 std::vector<Eigen::Vector3f> &points,
-                                std::vector<Eigen::Vector2f> &uvs) = 0;
+                                std::vector<Eigen::Vector3f> &normals,
+                                std::vector<Eigen::Vector3f> &colors) = 0;
+
+protected:
+    void SetLighting(const ViewControl &view, const RenderOption &option);
 
 protected:
     GLuint vertex_position_;
-    GLuint vertex_uv_;
-    GLuint texture_;
+    GLuint vertex_position_buffer_;
+    GLuint vertex_color_;
+    GLuint vertex_color_buffer_;
+    GLuint vertex_normal_;
+    GLuint vertex_normal_buffer_;
     GLuint MVP_;
+    GLuint V_;
+    GLuint M_;
+    GLuint light_position_world_;
+    GLuint light_color_;
+    GLuint light_diffuse_power_;
+    GLuint light_specular_power_;
+    GLuint light_specular_shininess_;
+    GLuint light_ambient_;
 
-    int num_materials_;
-    std::vector<int> array_offsets_;
-    std::vector<GLsizei> draw_array_sizes_;
-
-    std::vector<GLuint> vertex_position_buffers_;
-    std::vector<GLuint> vertex_uv_buffers_;
-    std::vector<GLuint> texture_buffers_;
+    // At most support 4 lights
+    GLHelper::GLMatrix4f light_position_world_data_;
+    GLHelper::GLMatrix4f light_color_data_;
+    GLHelper::GLVector4f light_diffuse_power_data_;
+    GLHelper::GLVector4f light_specular_power_data_;
+    GLHelper::GLVector4f light_specular_shininess_data_;
+    GLHelper::GLVector4f light_ambient_data_;
 };
 
-class TextureSimpleShaderForTriangleMesh : public TextureSimpleShader {
+class PhongShaderForPointCloud : public PhongShader {
 public:
-    TextureSimpleShaderForTriangleMesh()
-        : TextureSimpleShader("TextureSimpleShaderForTriangleMesh") {}
+    PhongShaderForPointCloud() : PhongShader("PhongShaderForPointCloud") {}
 
 protected:
     bool PrepareRendering(const geometry::Geometry &geometry,
@@ -94,7 +106,24 @@ protected:
                         const RenderOption &option,
                         const ViewControl &view,
                         std::vector<Eigen::Vector3f> &points,
-                        std::vector<Eigen::Vector2f> &uvs) final;
+                        std::vector<Eigen::Vector3f> &normals,
+                        std::vector<Eigen::Vector3f> &colors) final;
+};
+
+class PhongShaderForTriangleMesh : public PhongShader {
+public:
+    PhongShaderForTriangleMesh() : PhongShader("PhongShaderForTriangleMesh") {}
+
+protected:
+    bool PrepareRendering(const geometry::Geometry &geometry,
+                          const RenderOption &option,
+                          const ViewControl &view) final;
+    bool PrepareBinding(const geometry::Geometry &geometry,
+                        const RenderOption &option,
+                        const ViewControl &view,
+                        std::vector<Eigen::Vector3f> &points,
+                        std::vector<Eigen::Vector3f> &normals,
+                        std::vector<Eigen::Vector3f> &colors) final;
 };
 
 }  // namespace glsl
