+++ cpp/open3d/visualization/shader/RGBDImageShader.cpp
@@ -30,7 +30,7 @@
 #include <algorithm>
 
 #include "open3d/geometry/Image.h"
-#include "open3d/visualization/Shader/Shader.h"
+#include "open3d/visualization/shader/Shader.h"
 #include "open3d/visualization/utility/ColorMap.h"
 
 namespace open3d {
