+++ cpp/open3d/visualization/shader/Simple2DShader.h
@@ -29,19 +29,19 @@
 #include <Eigen/Core>
 #include <vector>
 
-#include "open3d/visualization/Shader/ShaderWrapper.h"
+#include "open3d/visualization/shader/ShaderWrapper.h"
 
 namespace open3d {
 namespace visualization {
 
 namespace glsl {
 
-class PhongShader : public ShaderWrapper {
+class Simple2DShader : public ShaderWrapper {
 public:
-    ~PhongShader() override { Release(); }
+    ~Simple2DShader() override { Release(); }
 
 protected:
-    PhongShader(const std::string &name) : ShaderWrapper(name) { Compile(); }
+    Simple2DShader(const std::string &name) : ShaderWrapper(name) { Compile(); }
 
 protected:
     bool Compile() final;
@@ -62,41 +62,19 @@ protected:
                                 const RenderOption &option,
                                 const ViewControl &view,
                                 std::vector<Eigen::Vector3f> &points,
-                                std::vector<Eigen::Vector3f> &normals,
                                 std::vector<Eigen::Vector3f> &colors) = 0;
-
-protected:
-    void SetLighting(const ViewControl &view, const RenderOption &option);
 
 protected:
     GLuint vertex_position_;
     GLuint vertex_position_buffer_;
     GLuint vertex_color_;
     GLuint vertex_color_buffer_;
-    GLuint vertex_normal_;
-    GLuint vertex_normal_buffer_;
-    GLuint MVP_;
-    GLuint V_;
-    GLuint M_;
-    GLuint light_position_world_;
-    GLuint light_color_;
-    GLuint light_diffuse_power_;
-    GLuint light_specular_power_;
-    GLuint light_specular_shininess_;
-    GLuint light_ambient_;
-
-    // At most support 4 lights
-    GLHelper::GLMatrix4f light_position_world_data_;
-    GLHelper::GLMatrix4f light_color_data_;
-    GLHelper::GLVector4f light_diffuse_power_data_;
-    GLHelper::GLVector4f light_specular_power_data_;
-    GLHelper::GLVector4f light_specular_shininess_data_;
-    GLHelper::GLVector4f light_ambient_data_;
 };
 
-class PhongShaderForPointCloud : public PhongShader {
+class Simple2DShaderForSelectionPolygon : public Simple2DShader {
 public:
-    PhongShaderForPointCloud() : PhongShader("PhongShaderForPointCloud") {}
+    Simple2DShaderForSelectionPolygon()
+        : Simple2DShader("Simple2DShaderForSelectionPolygon") {}
 
 protected:
     bool PrepareRendering(const geometry::Geometry &geometry,
@@ -106,23 +84,6 @@ protected:
                         const RenderOption &option,
                         const ViewControl &view,
                         std::vector<Eigen::Vector3f> &points,
-                        std::vector<Eigen::Vector3f> &normals,
-                        std::vector<Eigen::Vector3f> &colors) final;
-};
-
-class PhongShaderForTriangleMesh : public PhongShader {
-public:
-    PhongShaderForTriangleMesh() : PhongShader("PhongShaderForTriangleMesh") {}
-
-protected:
-    bool PrepareRendering(const geometry::Geometry &geometry,
-                          const RenderOption &option,
-                          const ViewControl &view) final;
-    bool PrepareBinding(const geometry::Geometry &geometry,
-                        const RenderOption &option,
-                        const ViewControl &view,
-                        std::vector<Eigen::Vector3f> &points,
-                        std::vector<Eigen::Vector3f> &normals,
                         std::vector<Eigen::Vector3f> &colors) final;
 };
