+++ cpp/open3d/visualization/shader/TexturePhongShader.h
@@ -29,19 +29,19 @@
 #include <Eigen/Core>
 #include <vector>
 
-#include "open3d/visualization/Shader/ShaderWrapper.h"
+#include "open3d/visualization/shader/ShaderWrapper.h"
 
 namespace open3d {
 namespace visualization {
 
 namespace glsl {
 
-class SimpleBlackShader : public ShaderWrapper {
+class TexturePhongShader : public ShaderWrapper {
 public:
-    ~SimpleBlackShader() override { Release(); }
+    ~TexturePhongShader() override { Release(); }
 
 protected:
-    SimpleBlackShader(const std::string &name) : ShaderWrapper(name) {
+    TexturePhongShader(const std::string &name) : ShaderWrapper(name) {
         Compile();
     }
 
@@ -63,18 +63,51 @@ protected:
     virtual bool PrepareBinding(const geometry::Geometry &geometry,
                                 const RenderOption &option,
                                 const ViewControl &view,
-                                std::vector<Eigen::Vector3f> &points) = 0;
+                                std::vector<Eigen::Vector3f> &points,
+                                std::vector<Eigen::Vector3f> &normals,
+                                std::vector<Eigen::Vector2f> &uvs) = 0;
+
+protected:
+    void SetLighting(const ViewControl &view, const RenderOption &option);
 
 protected:
     GLuint vertex_position_;
-    GLuint vertex_position_buffer_;
+    GLuint vertex_uv_;
+    GLuint vertex_normal_;
     GLuint MVP_;
+    GLuint V_;
+    GLuint M_;
+    GLuint light_position_world_;
+    GLuint light_color_;
+    GLuint light_diffuse_power_;
+    GLuint light_specular_power_;
+    GLuint light_specular_shininess_;
+    GLuint light_ambient_;
+
+    GLuint diffuse_texture_;
+
+    int num_materials_;
+    std::vector<int> array_offsets_;
+    std::vector<GLsizei> draw_array_sizes_;
+
+    std::vector<GLuint> vertex_position_buffers_;
+    std::vector<GLuint> vertex_uv_buffers_;
+    std::vector<GLuint> vertex_normal_buffers_;
+    std::vector<GLuint> diffuse_texture_buffers_;
+
+    // At most support 4 lights
+    GLHelper::GLMatrix4f light_position_world_data_;
+    GLHelper::GLMatrix4f light_color_data_;
+    GLHelper::GLVector4f light_diffuse_power_data_;
+    GLHelper::GLVector4f light_specular_power_data_;
+    GLHelper::GLVector4f light_specular_shininess_data_;
+    GLHelper::GLVector4f light_ambient_data_;
 };
 
-class SimpleBlackShaderForPointCloudNormal : public SimpleBlackShader {
+class TexturePhongShaderForTriangleMesh : public TexturePhongShader {
 public:
-    SimpleBlackShaderForPointCloudNormal()
-        : SimpleBlackShader("SimpleBlackShaderForPointCloudNormal") {}
+    TexturePhongShaderForTriangleMesh()
+        : TexturePhongShader("TexturePhongShaderForTriangleMesh") {}
 
 protected:
     bool PrepareRendering(const geometry::Geometry &geometry,
@@ -83,22 +116,9 @@ protected:
     bool PrepareBinding(const geometry::Geometry &geometry,
                         const RenderOption &option,
                         const ViewControl &view,
-                        std::vector<Eigen::Vector3f> &points) final;
-};
-
-class SimpleBlackShaderForTriangleMeshWireFrame : public SimpleBlackShader {
-public:
-    SimpleBlackShaderForTriangleMeshWireFrame()
-        : SimpleBlackShader("SimpleBlackShaderForTriangleMeshWireFrame") {}
-
-protected:
-    bool PrepareRendering(const geometry::Geometry &geometry,
-                          const RenderOption &option,
-                          const ViewControl &view) final;
-    bool PrepareBinding(const geometry::Geometry &geometry,
-                        const RenderOption &option,
-                        const ViewControl &view,
-                        std::vector<Eigen::Vector3f> &points) final;
+                        std::vector<Eigen::Vector3f> &points,
+                        std::vector<Eigen::Vector3f> &normals,
+                        std::vector<Eigen::Vector2f> &uvs) final;
 };
 
 }  // namespace glsl
